Fix precision and signedness of integer literal types in Java. Resolves #120.

+++ hydra-java/src/main/java/hydra/lib/equality/EqualInt8.java
@@ -5,16 +5,16 @@ import hydra.lib.PrimitiveType;
 import java.util.function.Function;
 
 
-public class EqualInt8 extends EqualityFunction<Short> {
+public class EqualInt8 extends EqualityFunction<Byte> {
     public EqualInt8() {
         super(PrimitiveType.int8(), Relation.EQUALS);
     }
 
-    public static Function<Short, Boolean> apply(Short second) {
+    public static Function<Byte, Boolean> apply(Byte second) {
         return first -> apply(first, second);
     }
 
-    public static Boolean apply(Short first, Short second) {
+    public static Boolean apply(Byte first, Byte second) {
         return first.equals(second);
     }
 }
